community/requirements.txt
pandas
playwright
psycopg2-binary